+++ src/main/java/com/project/cheerha/common/config/ElasticsearchConfig.java
@@ -19,7 +19,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
 
 @Configuration
-@EnableElasticsearchRepositories(basePackages = "com.project.cheerha.domain.jobopening.elasticrepository")
+@EnableElasticsearchRepositories(basePackages = "com.project.cheerha.domain.jobopening.elasticsearch")
 public class ElasticsearchConfig {
 
     @Value("${elasticsearch.uris}")
